fetch-market/requirements.txt
requests>=2.28.2
pandas>=2.0.3
python-dateutil>=2.9.0.post0
tqdm>=4.66.4
python-dotenv>=1.0.1 